jeni insert

+++ Jdbc/Reflection/src/main/java/ro/teamnet/zth/api/em/EntityManager.java
@@ -7,7 +7,7 @@ import java.util.List;
  */
 public interface EntityManager {
 
-    Integer getNextIdVal(String tableName, String columnIdName);
+    Long getNextIdVal(String tableName, String columnIdName);
     <T> T findById(Class<T> entityClass, Long id);
     <T> Object insert(T entity);
     <T> List<T> findAll(Class<T> entityClass);
